requirements.txt
beanie==1.29.0
fastapi==0.115.8
uvicorn==0.34.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.2.1
httpx==0.28.0
pydantic[email]
openai==1.61.1
python-dotenv
